* Peer Profiles - Preparation for using bonuses:       - Use CapacityBonus rather than ReachablilityBonus in the Capacity calculation       - Persist CapacityBonus rather than ReachabilityBonus       - Include SpeedBonus in the Speed calculation       - Prevent negative values in Speed and Capacity when using bonuses       - Clean up SpeedCalculator.java

+++ router/java/src/net/i2p/router/peermanager/PeerProfile.java
@@ -174,7 +174,7 @@ public class PeerProfile {
      * written to disk to affect how the algorithm ranks capacity.  Negative values are
      * penalties
      */
-    public double getCapacityBonus() { return _capacityBonus; }
+    public long getCapacityBonus() { return _capacityBonus; }
     public void setCapacityBonus(long bonus) { _capacityBonus = bonus; }
     
     /**
